Move object.go to dedicated package to prevent import cycle
